Move shared utils method to separate folder
Shared utils method for unit and integration tests are moved to
tests/helpers_utils folder.

CLOUDBLD-11124
