+++ tests/unit/package_managers/test_gomod.py
@@ -39,7 +39,7 @@ from cachi2.core.package_managers.gomod import (
     fetch_gomod_source,
 )
 from cachi2.core.packages_data import _package_sort_key
-from tests.unit.package_managers.helper_utils import assert_directories_equal, write_file_tree
+from tests.common_utils import assert_directories_equal, write_file_tree
 
 
 def setup_module():
